feat: Draft upgrade to Spring Boot 3 is implemented. Version is changed to 1.2.69-SNAPSHOT. Super-linter checkstyle config is renamed.

+++ atp-auth-models/src/main/java/org/qubership/atp/auth/springbootstarter/feign/exception/FeignClientException.java
@@ -68,7 +68,7 @@ public class FeignClientException extends RetryableException {
                                 final Request.HttpMethod httpMethod,
                                 final Map<String, Collection<String>> headers,
                                 final Request request) {
-        super(status, String.format("%d %s", status, errorMessage), httpMethod, null, request);
+        super(status, String.format("%d %s", status, errorMessage), httpMethod, (Long)null, request);
         this.status = status;
         this.errorMessage = errorMessage;
         this.headers = headers;
